simplify sample app
